requirements.txt
Scrapy==2.11.0
mysql-connector-python==8.2.0
boto3==1.34.0
requests==2.31.0
python-dotenv==1.0.0